# scan-engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scan-to-map-studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/groups/{name}": {
      "get": {
        "summary": "Api Group",
        "operationId": "api_group_api_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Api Delete Group",
        "operationId": "api_delete_group_api_groups__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/prepare": {
      "post": {
        "summary": "Api Prepare",
        "operationId": "api_prepare_api_groups__name__prepare_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/upload": {
      "post": {
        "summary": "Api Upload Group",
        "operationId": "api_upload_group_api_groups_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_upload_group_api_groups_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/workspace": {
      "get": {
        "summary": "Api Workspace",
        "description": "Slices + alignment + metrics + floor images as JSON (Fleet Studio native workspace).",
        "operationId": "api_workspace_api_groups__name__workspace_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{name}/merged.slicemap.json": {
      "get": {
        "summary": "Api Merged Slicemap",
        "operationId": "api_merged_slicemap_api_groups__name__merged_slicemap_json_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlicemapV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.floor.json": {
      "get": {
        "summary": "Api Merged Floor Json",
        "operationId": "api_merged_floor_json_api_groups__name__merged_floor_json_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/alignment": {
      "get": {
        "summary": "Api Get Alignment",
        "operationId": "api_get_alignment_api_groups__name__alignment_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupAlignmentDoc"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Api Put Alignment",
        "operationId": "api_put_alignment_api_groups__name__alignment_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupAlignmentDoc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveAlignmentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/alignment/history": {
      "get": {
        "summary": "Api Alignment History",
        "operationId": "api_alignment_history_api_groups__name__alignment_history_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlignmentHistoryEntry"
                  },
                  "title": "Response Api Alignment History Api Groups  Name  Alignment History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/alignment/history/{timestamp}": {
      "get": {
        "summary": "Api Alignment History Entry",
        "operationId": "api_alignment_history_entry_api_groups__name__alignment_history__timestamp__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupAlignmentDoc"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/icp": {
      "post": {
        "summary": "Api Icp",
        "operationId": "api_icp_api_groups__name__icp_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcpResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.png": {
      "get": {
        "summary": "Api Merged Png",
        "operationId": "api_merged_png_api_groups__name__merged_png_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.floor.png": {
      "get": {
        "summary": "Api Merged Floor Png",
        "operationId": "api_merged_floor_png_api_groups__name__merged_floor_png_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{name}": {
      "get": {
        "summary": "Group Workspace",
        "operationId": "group_workspace_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlignmentHistoryEntry": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "scans": {
            "type": "integer",
            "title": "Scans"
          },
          "approved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "scans",
          "approved"
        ],
        "title": "AlignmentHistoryEntry",
        "description": "One past save, newest first (GET /api/groups/{name}/alignment/history) -- a snapshot\ntaken right before that save overwrote group_alignment.json, so it's always restorable."
      },
...
      "Body_api_upload_group_api_groups_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_upload_group_api_groups_upload_post"
      },
      "FloorMeta": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format"
          },
          "resolution": {
            "type": "number",
            "title": "Resolution"
          },
          "origin": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Origin"
          },
          "width_px": {
            "type": "integer",
            "title": "Width Px"
          },
          "height_px": {
            "type": "integer",
            "title": "Height Px"
          },
          "row0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row0"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "format",
          "resolution",
          "origin",
          "width_px",
          "height_px"
        ],
        "title": "FloorMeta",
        "description": "merged.floor.json (== studio.floorplan.floor_meta)."
      },
...
      "GroupAlignmentDoc": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format",
            "default": "scan-group-alignment-v1"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "up_axis_convention": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Up Axis Convention"
          },
          "alignments": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AlignmentEntry"
            },
            "type": "object",
            "title": "Alignments"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "reference",
          "alignments"
        ],
        "title": "GroupAlignmentDoc",
        "description": "group_alignment.json (scan-group-alignment-v1)."
      },
      "GroupStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dir": {
            "type": "string",
            "title": "Dir"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "scans": {
            "items": {
              "$ref": "#/components/schemas/ScanStatus"
            },
            "type": "array",
            "title": "Scans"
          },
          "has_alignment": {
            "type": "boolean",
            "title": "Has Alignment"
          },
          "has_merged": {
            "type": "boolean",
            "title": "Has Merged"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "has_floor": {
            "type": "boolean",
            "title": "Has Floor",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "dir",
          "reference",
          "scans",
          "has_alignment",
          "has_merged",
          "ready"
        ],
        "title": "GroupStatus"
      },
      "GroupUploadResult": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "scans": {
            "type": "integer",
            "title": "Scans"
          }
        },
        "type": "object",
        "required": [
          "status",
          "group",
          "url",
          "scans"
        ],
        "title": "GroupUploadResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SlicemapV1": {
        "properties": {
          "format": {
            "type": "string",
            "const": "slicemap-v1",
            "title": "Format"
          },
          "z": {
            "type": "number",
            "title": "Z"
          },
          "band": {
            "type": "number",
            "title": "Band"
          },
          "resolution": {
            "type": "number",
            "title": "Resolution"
          },
          "origin": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Origin"
          },
          "cols": {
            "type": "integer",
            "title": "Cols"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "data": {
            "type": "string",
            "title": "Data"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SlicemapSource"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "format",
          "z",
          "band",
          "resolution",
          "origin",
          "cols",
          "rows",
          "data"
        ],
        "title": "SlicemapV1",
        "description": "slicemap-v1 file (merged.slicemap.json): base64 uint8 codes 0 unknown / 1 free / 2 furniture / 3 wall."
      },
...
      "WorkspacePayload": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerPayload"
            },
            "type": "array",
            "title": "Layers"
          },
          "gates": {
            "$ref": "#/components/schemas/WorkspaceGates"
          },
          "api": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceApi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "group",
          "reference",
          "layers",
          "gates"
        ],
        "title": "WorkspacePayload",
        "description": "GET /api/groups/{name}/workspace (== studio.align_workspace_html.workspace_payload)."
      }
    }
  }
}